requirements.txt
Django == 3.1      
django-json-response == 1.1.5         
pip==20.0.2   
pkg-resources==0.0.0    
pytz==2021.1   
requests == 2.25.1   
setuptools == 44.0.0   
sqlparse==0.4.1    
stripe==2.55.2   
urllib3==1.26.3